feat(currency): add formatCurrencyValue utility for consistent currency formatting
