+++ src/utils/currency-configs.ts
@@ -31,3 +31,37 @@ export const CURRENCY_CONFIGS: Record<Currency, CurrencyConfig> = {
   ETH_FUEL: CRYPTO_CONFIG,
   ETH: CRYPTO_CONFIG,
 };
+
+export const formatCurrencyValue = (
+  value: string,
+  config: CurrencyConfig,
+  includeLeadingZero: boolean,
+) => {
+  const [integerPart, decimalPart] = value.split(config.decimalSeparator);
+  let decimal = decimalPart || '';
+  let integer = integerPart || '';
+
+  if (config.thousandsSeparator) {
+    integer = integer.replace(
+      /\B(?=(\d{3})+(?!\d))/g,
+      config.thousandsSeparator,
+    );
+  }
+
+  if (includeLeadingZero) {
+    decimal = decimal.padEnd(config.decimalScale, '0');
+  }
+
+  if (!includeLeadingZero) {
+    const endsWithDecimal = value.endsWith(config.decimalSeparator);
+    if (decimal.length === 0) {
+      if (endsWithDecimal) {
+        return `${integer}${config.decimalSeparator}`;
+      }
+      return integer; // no decimal typed
+    }
+    return `${integer}${config.decimalSeparator}${decimal}`;
+  }
+
+  return `${integer}${config.decimalSeparator}${decimal}`;
+};
